pyproject.toml
[project]
name = "playbookgen"
version = "0.1.0"
dependencies = [
    "pathspec",
    "openai",
    "python-dotenv",
    "playwright",
    "PyYAML"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
